# FP.dev/contracts/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DoSelect.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5126/"
    }
  ],
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfOrderSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfOrderSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfOrderSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/orders/{id}/actions/cancel": {
      "post": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CancelOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/support-tickets/{id}": {
      "get": {
        "tags": [
          "AdminSupportTickets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSupportTicketDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSupportTicketDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSupportTicketDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/support-tickets": {
      "get": {
        "tags": [
          "SupportTickets"
        ],
        "parameters": [
          {
            "name": "Statuses",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SupportTicketStatus"
              }
            }
          },
          {
            "name": "Category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SupportTicketCategory"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfSupportTicketSummaryDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfSupportTicketSummaryDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultOfSupportTicketSummaryDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SupportTickets"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/support-tickets/{id}": {
      "get": {
        "tags": [
          "SupportTickets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/support-tickets/{id}/messages": {
      "post": {
        "tags": [
          "SupportTickets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportMessageRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportMessageRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/support-tickets/{id}/attachments": {
      "post": {
        "tags": [
          "SupportTickets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupportAttachmentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportAttachmentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportAttachmentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/support-tickets/{id}/actions/cancel": {
      "post": {
        "tags": [
          "SupportTickets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSupportTicketRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSupportTicketRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CancelSupportTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupportTicketDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminSupportTicketDetailDto": {
        "required": [
          "publicId",
          "ticketNumber",
          "category",
          "subject",
          "status",
          "priority",
          "orderPublicId",
          "assignee",
          "createdAtUtc",
          "lastActivityAtUtc",
          "firstResponseDueAtUtc",
          "resolutionDueAtUtc",
          "isOverdue",
          "firstHumanResponseAtUtc",
          "resolvedAtUtc",
          "closedAtUtc",
          "reopenCount",
          "availableActions",
          "rowVersion",
          "messages"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "format": "uuid"
          },
          "ticketNumber": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/SupportTicketCategory"
          },
          "subject": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SupportTicketStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority"
          },
          "orderPublicId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "assignee": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AdminAssigneeSummaryDto"
              }
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "lastActivityAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "firstResponseDueAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "resolutionDueAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "isOverdue": {
            "type": "boolean"
          },
          "firstHumanResponseAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "resolvedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "closedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reopenCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "availableActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSupportMessageDto"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportAttachmentDto"
            }
          }
        }
      },
...
      "CancelOrderRequest": {
        "required": [
          "reasonCode",
          "note",
          "orderRowVersion"
        ],
        "type": "object",
        "properties": {
          "reasonCode": {
            "type": "string"
          },
          "note": {
            "type": [
              "null",
              "string"
            ]
          },
          "orderRowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "CreateSupportTicketRequest": {
        "required": [
          "category",
          "subject",
          "message"
        ],
        "type": "object",
        "properties": {
          "category": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SupportTicketCategory"
              }
            ]
          },
          "subject": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "message": {
            "maxLength": 4000,
            "minLength": 1,
            "type": "string"
          },
          "orderPublicId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "OrderDto": {
        "required": [
          "publicId",
          "orderNumber",
          "orderStatus",
          "paymentStatus",
          "fulfillmentStatus",
          "assemblyStatus",
          "orderRefundStatus",
          "items",
          "recipient",
          "amounts",
          "paymentDueAtUtc",
          "confirmedAtUtc",
          "paidAtUtc",
          "shippedAtUtc",
          "deliveredAtUtc",
          "completedAtUtc",
          "cancelledAtUtc",
          "returnRequestDeadlineUtc",
          "availableActions",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "format": "uuid"
          },
          "orderNumber": {
            "type": "string"
          },
          "orderStatus": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "paymentStatus": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "fulfillmentStatus": {
            "$ref": "#/components/schemas/FulfillmentStatus"
          },
          "assemblyStatus": {
            "$ref": "#/components/schemas/AssemblyStatus"
          },
          "orderRefundStatus": {
            "$ref": "#/components/schemas/OrderRefundStatus"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemDto"
            }
          },
          "recipient": {
            "$ref": "#/components/schemas/OrderRecipientSummaryDto"
          },
          "amounts": {
            "$ref": "#/components/schemas/OrderAmountsDto"
          },
          "paymentDueAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "confirmedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "paidAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "shippedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deliveredAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "completedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "cancelledAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "returnRequestDeadlineUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "availableActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
...
      "PageResultOfOrderSummaryDto": {
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderSummaryDto"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "PageResultOfSupportTicketSummaryDto": {
        "required": [
          "items",
          "pageNumber",
          "pageSize",
          "totalCount"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportTicketSummaryDto"
            }
          },
          "pageNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "totalPages": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "SupportTicketCategory": {
        "enum": [
          "order",
          "payment",
          "logistics",
          "productWarranty",
          "returnHelp",
          "account",
          "other"
        ]
      },
      "SupportTicketDto": {
        "required": [
          "publicId",
          "ticketNumber",
          "category",
          "subject",
          "status",
          "priority",
          "orderPublicId",
          "createdAtUtc",
          "lastActivityAtUtc",
          "firstResponseDueAtUtc",
          "resolutionDueAtUtc",
          "isOverdue",
          "firstHumanResponseAtUtc",
          "resolvedAtUtc",
          "closedAtUtc",
          "reopenCount",
          "availableActions",
          "messages",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "publicId": {
            "type": "string",
            "format": "uuid"
          },
          "ticketNumber": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/SupportTicketCategory"
          },
          "subject": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SupportTicketStatus"
          },
          "priority": {
            "$ref": "#/components/schemas/CasePriority"
          },
          "orderPublicId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "lastActivityAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "firstResponseDueAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "resolutionDueAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "isOverdue": {
            "type": "boolean"
          },
          "firstHumanResponseAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "resolvedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "closedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reopenCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "availableActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportMessageDto"
            }
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupportAttachmentDto"
            }
          }
        }
      },
      "SupportTicketStatus": {
        "enum": [
          "open",
          "assigned",
          "inProgress",
          "waitingForCustomer",
          "waitingForInternal",
          "resolved",
          "closed",
          "cancelled"
        ]
      },
...
    }
  },
...
}